refactor: mvc 패턴 적용 해제
-MVC 패턴을 사용 하지 않고 객체 단위로 구현할 수 있도록 변경
