--- src/main/java/blackjack/domain/card/CardShape.java
@@ -1,18 +1,0 @@
-package blackjack.domain.card;
-
-public enum CardShape {
-    DIAMOND("다이아몬드"),
-    HEART("하트"),
-    CLOVER("클로버"),
-    SPADE("스페이드");
-
-    private final String displayName;
-
-    CardShape(final String displayName) {
-        this.displayName = displayName;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-}
