--- src/main/java/blackjack/domain/card/CardShuffler.java
@@ -1,12 +1,0 @@
-package blackjack.domain.card;
-
-import java.util.Collections;
-import java.util.List;
-
-public class CardShuffler implements Shuffler {
-    @Override
-    public List<Card> shuffle(final List<Card> cards) {
-        Collections.shuffle(cards);
-        return cards;
-    }
-}
