--- src/main/java/blackjack/domain/card/CardType.java
@@ -1,33 +1,0 @@
-package blackjack.domain.card;
-
-public enum CardType {
-    TWO(2, "2"),
-    THREE(3, "3"),
-    FOUR(4, "4"),
-    FIVE(5, "5"),
-    SIX(6, "6"),
-    SEVEN(7, "7"),
-    EIGHT(8, "8"),
-    NINE(9, "9"),
-    TEN(10, "10"),
-    KING(10, "K"),
-    QUEEN(10, "Q"),
-    JACK(10, "J"),
-    ACE(11, "A"); // 에이스의 값은 최대 값이 11로 정한다.
-
-    private final int value;
-    private final String displayName;
-
-    CardType(final int value, final String displayName) {
-        this.value = value;
-        this.displayName = displayName;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-}
