--- src/main/java/blackjack/domain/card/Shuffler.java
@@ -1,7 +1,0 @@
-package blackjack.domain.card;
-
-import java.util.List;
-
-public interface Shuffler {
-    List<Card> shuffle(List<Card> cards);
-}
