--- src/main/java/blackjack/domain/gambler/Dealer.java
@@ -1,17 +1,0 @@
-package blackjack.domain.gambler;
-
-import blackjack.domain.card.Card;
-import java.util.List;
-
-public class Dealer extends Gambler {
-    public Dealer() {
-        super(Name.getDealerName());
-    }
-
-    @Override
-    public List<Card> getInitialCards() {
-        List<Card> cards = getCards();
-        Card firstCard = cards.getFirst();
-        return List.of(firstCard);
-    }
-}
