--- src/main/java/blackjack/domain/gambler/Player.java
@@ -1,15 +1,0 @@
-package blackjack.domain.gambler;
-
-import blackjack.domain.card.Card;
-import java.util.List;
-
-public class Player extends Gambler {
-    public Player(final Name name) {
-        super(name);
-    }
-
-    @Override
-    public List<Card> getInitialCards() {
-        return getCards();
-    }
-}
