+++ src/main/java/blackjack/object/card/Shuffler.java
@@ -1,0 +1,7 @@
+package blackjack.object.card;
+
+import java.util.List;
+
+public interface Shuffler {
+    List<Card> shuffle(List<Card> cards);
+}
